+++ apps/gym/src/app/(auth)/layout.tsx
@@ -1,0 +1,7 @@
+export const metadata = {
+  title: 'Gym App',
+}
+
+export default function AuthLayout({ children }: { children: React.ReactNode }) {
+  return <div>{children}</div>
+}
