+++ apps/gym/src/app/(auth)/login/page.tsx
@@ -1,0 +1,7 @@
+import styles from './Login.module.css'
+
+const Login = () => {
+  return <div className={styles.container}>Login</div>
+}
+
+export default Login
