+++ apps/gym/src/app/(auth)/registration/page.tsx
@@ -1,0 +1,7 @@
+import styles from './Registration.module.css'
+
+const Registration = () => {
+  return <div className={styles.container}>Registration</div>
+}
+
+export default Registration
