+++ apps/gym/src/app/(user)/record/page.tsx
