+++ apps/gym/src/app/(user)/record/workout/page.tsx
