+++ apps/gym/src/app/(user)/splits/page.tsx
