+++ apps/gym/src/app/page.tsx
@@ -1,5 +1,16 @@
+'use client'
+
+import { Button } from '@gymapp/gymui/Button'
+import { useRouter } from 'next/navigation'
+
 const Home = () => {
-  return <div>This is a test for login</div>
+  const router = useRouter()
+  return (
+    <div>
+      <div>Home Page</div>
+      <Button.Primary onClick={() => router.push('/dashboard')}>Dashboard</Button.Primary>
+    </div>
+  )
 }
 
 export default Home
